fixed tests, fixed event error
