+++ src/js/level_choice.ts
@@ -11,20 +11,18 @@ let clicked = false
 levelChoicesDiv?.addEventListener('click', (event) => {
     event.preventDefault()
 
-    if (event.target instanceof HTMLElement) {
-        const target = event.target
+    const target = event.target as HTMLElement
 
-        if (target.closest('button')) {
-            console.log(target.dataset.level)
+    if (target.closest('button')) {
+        console.log(target.dataset.level)
 
-            levelChoiceButtons.forEach((levelButton) => {
-                levelButton.classList.remove('error')
-            })
+        levelChoiceButtons.forEach((levelButton) => {
+            levelButton.classList.remove('error')
+        })
 
-            localStorage.setItem('level', target.dataset.level!)
+        localStorage.setItem('level', target.dataset.level!)
 
-            clicked = true
-        }
+        clicked = true
     }
 })
